Move :bind-exwm into a separate file
Manageability is important when it comes to source code.

Since it's been taking over 50 lines of code to get to the first
`use-package' declaration, I've decided to separate :bind-exwm from
the rest of init-file.

This commit includes a lot of extra work on top of the new file. Many
of the previous "command generators" are now available through simple
changes of COMMAND's type.