Add an experimental alarm clock
It's not guaranteed to work at all.

The sole reason for its existence is this: the external `alarm-clock'
package simply starts one-off timers, and requires a few overrides to
work correctly.

The "experimental" package solves it all. It runs an alarm on
specified days of week, on specified time, with specified
ringtone. Dead simple.