beets: add lyrics plugin
It's almost completely useless because none of the music I listen to
has lyrics on Musixmatch.

Not only that, but the music that does have lyrics cannot be
recognised because of mismatching artist names and song titles (e.g.,
one is romaji, and what I have locally is kanji).

Musicbrainz and Musixmatch rarely agree with each other on this matter
because the latter uses the Spotify database for tagging music.